Move getOrCreateClassDecoratorStash to ClassStash
