+++ cocos/core/data/decorators/utils.ts
@@ -22,8 +22,6 @@
  THE SOFTWARE.
 */
 
-import type { ClassStash } from '../class-stash';
-
 export type Initializer = () => unknown;
 
 export type BabelPropertyDecoratorDescriptor = PropertyDescriptor & { initializer?: Initializer };
@@ -99,27 +97,4 @@ export function makeSmartClassDecorator<TArg> (
             };
         }
     }
-}
-
-const classDecoratorStashMap = new WeakMap<AnyFunction, ClassStash>();
-
-export function getOrCreateClassDecoratorStash (cls: AnyFunction, decoratorName?: string): ClassStash {
-    let stash = classDecoratorStashMap.get(cls);
-    if (!stash) {
-        stash = {};
-        classDecoratorStashMap.set(cls, stash);
-    }
-    return stash;
-}
-
-export function getClassDecoratorStash (cls: AnyFunction): ClassStash | undefined {
-    return classDecoratorStashMap.get(cls);
-}
-
-export function deleteClassDecoratorStash (cls: AnyFunction): void {
-    classDecoratorStashMap.delete(cls);
-}
-
-export function getSubDict<T, TKey extends keyof T> (obj: T, key: TKey): NonNullable<T[TKey]> {
-    return obj[key] as NonNullable<T[TKey]> || ((obj[key]) = {} as NonNullable<T[TKey]>);
 }
